Fix bug with gradient drawing opacity
The gradient and random buttons have an event handler attached that
calls the drawOnGrid function that change the draw type from gradient
to random colour and vice versa. This has unintentional functionality
in that by calling the function by the event handler attached to the
buttons, the opacity increment part of the drawOnGrid function is called
every button press.

This leads to the "mouseenter" event applying the colour or gradient as
many times as the button has been pressed rather than once as intended.
This also affects switching back to gradient from random. Gradient is
set at a higher value than the intended value 0.2 and increments
incorrectly.

The fix for this issue is to remove the drawOnGrid function call from
the buttons so the opacity is incremented correctly, only when mouseover
occurs and resets when switching draw colours.